Update dependencies
This commit updates the parent container images to the latest available
UBI and OpenJDK versions. NOTE: With UBI9, the `mvn
dependency:go-offline` command fails with a permission error. Adding
`USER root` just before excecuting the command works around this
limitation. This is ok, because the default `USER` is still used in the
final build stage in the Dockerfile.

This commit also updates the version of spring-boot to address several
security updates.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>